refactor: reorganize files for reduced redundancy
revert pre-commit.husky changes, use ReactDOMServer for HTML generation
