+++ package.json
@@ -20,6 +20,7 @@
     "serve": "gatsby serve",
     "clean": "gatsby clean",
     "test": "jest",
+    "format": "prettier \"**/*.{tsx,ts,md}\" --write",
     "test:types": "tsc --noEmit",
     "test:lint": "eslint \"**/*.{tsx,ts}\"",
     "test:format": "prettier \"**/*.{tsx,ts,md}\" --check",
